Extract Three.js custom utils to a separate file
