+++ cupLayoutTool/src/jsMain/kotlin/three/three.kt
@@ -1,7 +1,5 @@
 package three
 
-import dev.toolkt.geometry.Point3D
-import dev.toolkt.math.algebra.linear.vectors.Vector3
 import org.w3c.dom.HTMLCanvasElement
 import org.w3c.dom.HTMLElement
 import org.w3c.dom.extra.jsObject
@@ -280,26 +278,6 @@ external object THREE {
 
         fun invert()
     }
-}
-
-fun List<Point3D>.toBufferAttribute(): THREE.BufferAttribute = map { it.pointVector }.toBufferAttribute()
-
-fun List<Vector3>.toBufferAttribute(): THREE.BufferAttribute {
-    return THREE.BufferAttribute(
-        array = Float32Array(
-            this.flatMap { it.toList() }.toTypedArray(),
-        ),
-        itemSize = 3,
-    )
-}
-
-fun THREE.BufferGeometry.setPositionAttribute(
-    positions: List<Point3D>,
-) {
-    setAttribute(
-        name = "position",
-        attribute = positions.toBufferAttribute(),
-    )
 }
 
 val THREE.Object3D.worldPosition: THREE.Vector3
